refactor: FollowUserInfo 패키지 이동
