+++ user/src/main/kotlin/com/fx/user/adapter/security/JwtProvider.kt
@@ -2,8 +2,8 @@ package com.fx.user.adapter.security
 
 import com.fx.global.annotation.hexagonal.SecurityAdapter
 import com.fx.global.dto.UserRole
-import com.fx.user.application.out.TokenStoragePort
-import com.fx.user.application.out.JwtProviderPort
+import com.fx.user.application.out.security.TokenStoragePort
+import com.fx.user.application.out.security.JwtProviderPort
 import com.fx.user.domain.AuthenticatedUserInfo
 import com.fx.user.domain.TokenInfo
 import com.fx.user.exception.JwtException
